Change git log aliases to use Author instead of Committer name
Authors are the people from whom the commits originate from. Committer is simply
the person who makes the commit. You can see the difference if you perform a
rebase of commits on a branch which multiple people work on, because the commit
authors are the various people, but the committer becomes you.