remove omz's ssh-agent support, add ssh-agent forwarding to ssh config
Problem: we use the same dotfiles on all servers. ssh-agent forwarding does not
work well if ssh-agent exists, because in forwarding sshd acts as the agent.

To avoid this, we're removing ssh-agent setup from dotfiles, because it's not
necessary anyway - we'll rely on OS X and Ubuntu and Pagent's (windows, PuTTY
complement) built in prompts to have the identity added.

Most servers are already prepped to accept ssh-agent forwarding, so the real
magic that gives us forwarding is our `.ssh/config` file. Just put `ForwardAgent
yes` on any machines that you want to forward identities to.

Try to avoid placing `ForwardAgent` in the config's global settings.