add check to run ssh-add if identity not loaded
I mentioned we could rely on OS prompts to add identities, but it proved to be
more convenient to just ask at the start of any terminal session, rather than
wait for the first ssh attempt.

Also, this means my tmuxinator setups won't have the chance to ask for the
passphrase.