move to vimperator
Much faster than pentadactyl on any of my machines. I don't use pentadactyl to
its fullest customisability/power.

Goodbye, Pentadactyl - maybe if you get more resource efficient and not so
bloated. It's a bit of a downgrade, as Pentadactyl has...

- better tab numbering (don't know HOW the Vimperator devs can handle something
  so simple, so incompetently)

- better combined commandline/statusbar (URL visibility on Vimperator's sucks)

...but those can Probably be solved with Vimperator plugins and CSS styling.
With Vimperator's speed, it is a trade-off I'm willing to make.