move rainbarf to tmux folder
Does not belong in the powerline folder. Also, have discovered that Inconsolata
does not display the following correctly:

▁ ▂ ▃ ▄ ▅ ▆ ▇ █

https://github.com/creaktive/rainbarf/issues/10

This string can be used to see if a font properly supports the usage bars.
As it turns out, Menlo does (using the powerline-patched Meslo works well).

There is still the issue with Menlo where some of the red fonts are raised 1
pixel over the others.