Change prompt character, show navigation in vimperator
Surprisingly, the dash as a prompt character gives more of the same feel as the
multibyte "→" than the angle bracket does.

The navigation bar in firefox is now more useful, because the more recent
versions of firefox (30+?) no longer support the plugin toolbar underwhich
xdebug and proxy icons could live - this is all now built into the same toolbar
that navigation is under. Until we work out vimperator commands for those, we'll
leave the navigation bar for mouse activation.