Update theme git prompt
We finally reduce the number of calls to `git status` by overriding
`parse_git_dirty`. More variables are also given local scope, although not sure
if this improves anything. At the rate we are overriding omz library functions,
we should perhaps make the theme entirely self-sufficient, and then see if we
can move to prezto...