Add vim-pandoc for document creation from markdown
This allows a user to create `.pandoc` text files, and produce
same-named `.docx` files (default) on save or update. For writing
purposes, vim-pandoc + goyo/limelight is an excellent pairing!

Pandoc is capable of many other formats and can also integrate with
bibliography formats!

To learn the pandoc markdown syntax, refer to
http://pandoc.org/README.html#pandocs-markdown