Expand vimrc to be generally self-sufficient
We are updating the vimrc to contain all important settings and mappings
such that it can be dumped on most any machine and provide a comfortable
working environment. The vimrc should work well regardless of there
being additional settings or plugins. We also look to making it
increasingly compatible with various older or less feature-full versions
of vim, using conditionals.