Add `cu` alias and completion
`cu` can be used in any terminal emulator to talk to serial
devices/consoles/ports, instead of using a dedicated console application
like ZTerm. This is useful for managing cisco devices, for example.

Also got to learn about the zsh completion system. Now, any custom
completions that we do not want to turn into a plugin can be put in
`zsh/custom/completions`.