Simplify ssh-agent script, do not persist in sudo
We no longer support persisting an agent across `sudo`: it is needlessly
insecure and not actually used often. The main benefit of symlinking the
socket is now persisting across multiplexers.

The symlink is now moved to within a user's `.ssh` folder, which
would/should have strict permissions (700).

Of course, the socket itself is still public and so exploitable by any
concurrent user with root access - the standard warning to use
ssh agent forwarding sparingly/only-on-super-trusted-machines applies.