Initial migration of shell config
Shell config now caters to both bash and zsh.
Scripts common to bash and zsh have been split into a 'common' folder.
Install scripts have been updated to install bash and zsh separately.
Comprehensive configuration of zsh and bash has yet to be done, but the
basics are in place.