Tweak vim-pandoc settings, add slide creation
We make Goyo play nice with vim-pandoc's foldcolumn, default to odt
files (working on getting a good odt reference template). The `A`
formatting option is removed for better performance on larger paragraphs
in long documents (instead, we need to format it ourselves with `gqip`
or run ':FormatPandoc'). Side scrolloff is lowered to stop vim from
jumping left and right lots when terminal is fullscreen
e.g. when using it side by side with another window.