Remove more vim plugins and solarized theme
More pruning of the dotfiles. Some of the plugins have been moved to our
local extra-configs as context specific things.

We also take the `papercolor` theme out of the `vim-airline` repository,
as we really don't use any of the others. Current `:Light` colour
scheme is provisional, and subject to change later, when we find
something as fantastic as Neonwave/Suraken.

Sadly, the ultimate/lightest completion system is yet to be found.